Get all and rand links from firebase
