+++ server/db.js
@@ -1,5 +1,6 @@
 import dotenv from 'dotenv';
 import admin from 'firebase-admin';
+import { getRandomInt } from './utils.js';
 
 dotenv.config()
 
@@ -14,6 +15,24 @@ admin.initializeApp({
 
 // As an admin, the app has access to read and write all data, regardless of Security Rules
 const db = admin.database();
+
+export const getAllConcertLinks = async () => {
+  let concertLinks = [];
+  
+  const ref = db.ref("concerts/links/");
+  await ref.once("value", function(snapshot) {
+    concertLinks = snapshot.val();
+  });
+
+  return concertLinks;
+};
+
+export const getRandomConcert = async () => {
+  const concertLinks = await getAllConcertLinks();
+  const randConcertLink = concertLinks[getRandomInt(concertLinks.length - 1)];
+  
+  return  randConcertLink;
+}
 
 export default db;
